Start drafting an alternative version of Inversion
This new class doesn't execute directives, so it doesn't apply any
change to the objective function. The idea is that users intentionally
apply them, either manually or through directives. This way directives
can be more flexible (they can be whatever we want them to be). We can
even redefine the objective function, without having to modify in-place.