Incorporate directives into the `Inversion` class (#9)
Add a new `directives` argument to the `Inversion` class. The `__next__`
method calls all directives after the zero-th iteration. The directives
are defined as callables that should take the model and the iteration
number as arguments. Update the inversion constructor. Add notebook that
runs the linear regression using the constructor function. Update and
rerun the notebooks.