Use default log in Inversion
Use a default log in Inversion if `log=True`, so the user doesn't need
to build it and pass it. If `log=False`, then `Inversion.log` will be
`None`. Users can still pass their custom `InversionLog`. Move the first
call of `log.update()` to the `__next__` method, only executed in the
first iteration. This makes Inversion instantiation faster (no calls to
objective functions), and allows users to edit the log columns before
the inversion has started.