Add `update` method to conditions (#16)
Add a new `Condition.update` method that takes a model and allow us to
update the stopping criteria if needed. Implement `ModelChanged` and
`ObjectiveChanged` conditions that return true when the model and an
objective function change above a tolerance, respectively. Use a private
mixin class for the public ones. Add tests.