Allow ConjugateGradient__call__ to take preconditioner
Add an optional `preconditioner` argument to the
`ConjugateGradient.__call__()` method: it can be a dense or sparse
matrix, a `LinearOperator` or a callable that takes the model as
argument and returns a matrix or `LinearOperator`. By passing a callable
we can allow users to build the preconditioner dynamically. Add
a `kwargs` argument to the `Inversion` that will be passed to the
`optimizer`. This way users can pass a `preconditioner` argument to the
`Inversion`, which will be used by the minimizer.
a