Add support for preconditioners (#27)
Add a `hessian_diagonal` abstract method to `Objective`, implement it
for `Combo`, `Scaled`, and the `DataMisfit` and the regularizations. Add
a `utils.get_jacobi_preconditioner` function that builds a Jacobi
preconditioner from an objective function and a given model. Allow
`ConjugateGradient` to take a `preconditioner` argument in the
`__call__` method: it can be a predefined matrix (dense, sparse or
`LinearOperator`), or a callable that takes the `model` as argument and
returns a matrix (same types). Allow `Inversion` to take `kwargs` that
get passed to the `minimizer`. By passing `preconditioner` as a function
to the `Inversion` we can easily update the preconditioner on every
iteration. Add `preconditioner` and `update_preconditioner` arguments to
the inversion constructor function: the `preconditioner` can be just
`"jacobi"`, so the function will automatically build the Jacobi
preconditioner. Add some notebooks and update some of the current ones.

Fixes #20