Add `JacobiPreconditioner` class (#33)
The new class returns a callable object that can be used as
preconditioner: it generates a Jacobi preconditioner for a given
objective function. This class can be used to update the preconditioner
on every iteration through an inversion. Create a new
`preconditioner.py` file where the new class and the
`get_jacobi_preconditioner` function live. Simplify the inversion
constructor: Remove the `update_preconditioner` argument: if we pass
`"jacobi"` as preconditioner, it'll define a `JacobiPreconditioner` that
will get updated on each iteration. Users can easily define predefined
preconditioners instead, no need to add complexity with another argument
in the function. Update and rerun notebooks 11 and 12.