Update definition of `DataMisfit`s weights (#37)
Fix definition of weights in the data misfit. They follow a similar
definition to regularization weights: they are the square of the inverse
of uncertainties. The weights matrix is still a diagonal matrix with the
inverse of the uncertainties (square root of weights). Minor update to
docs of sensitivity weights function to make it clear it requires the
data weights matrix.