Add a `get_minimizer_callback` method to Log
The `InversionLog` now holds a `get_minimizer_callback` method that
returns a function that can be passed to `Minimizer`'s as the `callback`
argument. The `InversionLog` creates a new `MinimizerLog` every time the
`get_minimizer_callback` is called, appends it to a running list of
`minimizer_logs`, and returns the `update` method of the newly created
one. The `Inversion` then passes this callable to the `Minimizer` when
called. This way the logs of all minimization steps are being stored
inside the `InversionLog`.