Keep allocating the Rich table
It's best to build the Rich renderables and store them in memory rather
than rebuilding them every time the `__rich__` method is called. This is
to avoid async issues that might happen when the `Live` wants to update
the renderable, but some pieces of the code required to build it is
still running (filling the log for example). So it's better to always
have one renderable ready to be shown.